+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased/Snapshot]
 
+### Added
+-  Definition for a default input file directory structure
+-  Tarball utils to extract and compress files
+
 ### Changed
 - moved api docs to own branch incl. automated api-docs deployment